Created InputComponent
Abstract InputComponents, condenses code in AddComponentDialog,
increases readability and extensibility.  To add an InputComponent to
the AddComponentDialog, now create a class which extends the abstract
class InputComponent for the generic type required, and add to the
inputComponents hashmap in AddComponentDialog.  Will be reformed to use
Annotations in the future to ease InputComponent creation/assignment